pyproject.toml
[project]
name = "calendar-sync"
version = "0.1.0"
description = "Sync RSS feed events to Google Calendar using Claude"
requires-python = ">=3.11"
dependencies = [
    "feedparser>=6.0",
    "anthropic>=0.40",
    "google-api-python-client>=2.100",
    "google-auth>=2.20",
    "typer>=0.9",
    "pydantic>=2.0",
    "httpx>=0.25",
    "python-dotenv>=1.0",
]

[project.scripts]
calsync = "calendar_sync.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "boto3>=1.35",
    "ruff>=0.15.2",
    "ty>=0.0.17",
]
